requirements.txt
icnsutil==1.1.0
packaging==26.2
pillow==12.2.0
psutil==7.2.2
setuptools==82.0.1
wheel==0.47.0
